# example/petstore/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "petstore api",
        "title": "petstore api",
        "contact": {},
        "version": "1.0"
    },
    "paths": {
...
        "/store/inventory": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Returns a map of status codes to quantities\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "\"Returns pet inventories by status\"",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    }
                }
            }
        },
        "/store/order": {
            "post": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"Place a new order in the store\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "\"Place an order for a pet\"",
                "parameters": [
                    {
                        "description": "Place a new order in the store",
                        "name": "Order",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/order.OrderDTO"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/order.OrderDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid input"
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                }
            }
        },
        "/store/order/{orderId}": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"For valid response try integer IDs with value \u003c= 5 or \u003e 10. Other values will generate exceptions.\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "\"Find purchase order by ID\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of order that needs to be fetched",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation",
                        "schema": {
                            "$ref": "#/definitions/order.OrderDTO"
                        }
                    },
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            },
            "delete": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "\"For valid response try integer IDs with value \u003c 1000. Anything above 1000 or nonintegers will generate API errors\"",
                "consumes": [
                    "application/json"
                ],
                "tags": [
                    "store"
                ],
                "summary": "\"Delete purchase order by ID\"",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "ID of the order that needs to be deleted",
                        "name": "orderId",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "400": {
                        "description": "Invalid ID supplied"
                    },
                    "404": {
                        "description": "Order not found"
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "order.OrderDTO": {
            "type": "object",
            "properties": {
                "complete": {
                    "type": "boolean"
                },
                "id": {
                    "type": "integer"
                },
                "petId": {
                    "type": "integer"
                },
                "quantity": {
                    "type": "integer"
                },
                "shipDate": {
                    "type": "string"
                },
                "status": {
                    "description": "Order Status",
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}